refactor: move ensureNativeCompat to ensure-deps.mjs for automatic ABI protection (#203)
The ABI cache/rebuild logic was only in start.mjs (MCP server). Hooks
that load better-sqlite3 had no ABI protection, causing crashes when
the project's Node version differed from the one context-mode was
compiled against.

Now ensure-deps.mjs (imported by all hooks) runs both dependency
install AND ABI compat check automatically. No env var needed — it
just works. CONTEXT_MODE_NODE remains as an escape hatch.
