+++ start.mjs
@@ -1,53 +1,10 @@
 #!/usr/bin/env node
 import "./hooks/reexec-node.mjs";
 import { execSync } from "node:child_process";
-import { existsSync, copyFileSync, chmodSync, readFileSync, writeFileSync, readdirSync } from "node:fs";
+import { existsSync, chmodSync, readFileSync, writeFileSync, readdirSync } from "node:fs";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { homedir } from "node:os";
-import { createRequire } from "node:module";
-
-/**
- * ABI-aware native binary caching for better-sqlite3 (#148).
- * Users with mise/asdf may run concurrent sessions with different Node versions.
- * Each ABI needs its own compiled binary — cache them side-by-side.
- */
-function ensureNativeCompat(pluginRoot) {
-  try {
-    const abi = process.versions.modules;
-    const nativeDir = resolve(pluginRoot, "node_modules", "better-sqlite3", "build", "Release");
-    const binaryPath = resolve(nativeDir, "better_sqlite3.node");
-    const abiCachePath = resolve(nativeDir, `better_sqlite3.abi${abi}.node`);
-
-    if (!existsSync(nativeDir)) return;
-
-    if (existsSync(abiCachePath)) {
-      copyFileSync(abiCachePath, binaryPath);
-      return;
-    }
-
-    if (!existsSync(binaryPath)) return;
-
-    try {
-      const req = createRequire(resolve(pluginRoot, "package.json"));
-      req("better-sqlite3");
-      copyFileSync(binaryPath, abiCachePath);
-    } catch (probeErr) {
-      if (probeErr?.message?.includes("NODE_MODULE_VERSION")) {
-        execSync("npm rebuild better-sqlite3", {
-          cwd: pluginRoot,
-          stdio: "pipe",
-          timeout: 60000,
-        });
-        if (existsSync(binaryPath)) {
-          copyFileSync(binaryPath, abiCachePath);
-        }
-      }
-    }
-  } catch {
-    /* best effort — server will report the error on first DB access */
-  }
-}
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const originalCwd = process.cwd();
@@ -116,9 +73,10 @@ if (cacheMatch) {
   }
 }
 
-// Ensure native dependencies are available (shared with hooks via ensure-deps.mjs)
-import { ensureDeps } from "./hooks/ensure-deps.mjs";
-// ensure-deps handles better-sqlite3; also install pure-JS deps used by server
+// Ensure native dependencies + ABI compatibility (shared with hooks via ensure-deps.mjs)
+// ensure-deps handles better-sqlite3 install + ABI cache/rebuild automatically (#148, #203)
+import "./hooks/ensure-deps.mjs";
+// Also install pure-JS deps used by server
 for (const pkg of ["turndown", "turndown-plugin-gfm", "@mixmark-io/domino"]) {
   if (!existsSync(resolve(__dirname, "node_modules", pkg))) {
     try {
@@ -130,11 +88,6 @@ for (const pkg of ["turndown", "turndown-plugin-gfm", "@mixmark-io/domino"]) {
     } catch { /* best effort */ }
   }
 }
-
-// Ensure better-sqlite3 native binary matches current Node.js ABI (#148)
-// Users with mise/asdf may run concurrent sessions with different Node versions.
-// Each ABI needs its own compiled binary — cache them side-by-side.
-ensureNativeCompat(__dirname);
 
 // Self-heal: create CLI shim if cli.bundle.mjs is missing (marketplace installs)
 if (!existsSync(resolve(__dirname, "cli.bundle.mjs")) && existsSync(resolve(__dirname, "build", "cli.js"))) {
